refactor: extract scalar function OptimizeVTable
