+++ vortex-array/src/arrays/scalar_fn/vtable/mod.rs
@@ -271,6 +271,7 @@ impl Display for FakeEq<ArrayRef> {
 
 impl scalar_fn::ScalarFnVTable for ArrayExpr {
     type Options = FakeEq<ArrayRef>;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.array");
