+++ vortex-array/src/scalar_fn/fns/byte_length.rs
@@ -65,6 +65,7 @@ pub struct ByteLength;
 
 impl ScalarFnVTable for ByteLength {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.byte_length");
