+++ vortex-array/src/scalar_fn/fns/ext_storage.rs
@@ -27,6 +27,7 @@ pub struct ExtStorage;
 
 impl ScalarFnVTable for ExtStorage {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.ext.storage");
