+++ vortex-array/src/scalar_fn/fns/is_not_null.rs
@@ -41,6 +41,7 @@ impl IsNotNull {
 
 impl ScalarFnVTable for IsNotNull {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.is_not_null");
