+++ vortex-array/src/scalar_fn/fns/list_contains/mod.rs
@@ -64,6 +64,7 @@ impl ListContains {
 
 impl ScalarFnVTable for ListContains {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.list.contains");
