+++ vortex-array/src/scalar_fn/internal/row_count.rs
@@ -47,6 +47,7 @@ pub struct RowCount;
 
 impl ScalarFnVTable for RowCount {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.row_count");
