+++ vortex-layout/src/layouts/row_idx/expr.rs
@@ -25,6 +25,7 @@ pub struct RowIdx;
 
 impl ScalarFnVTable for RowIdx {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.row_idx");
