+++ vortex-spatial/src/scalar_fn/collect.rs
@@ -260,6 +260,7 @@ impl SpatialCollect {
 
 impl ScalarFnVTable for SpatialCollect {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.st.collect");
