+++ vortex-tensor/src/scalar_fns/cosine_similarity.rs
@@ -73,6 +73,7 @@ impl CosineSimilarity {
 
 impl ScalarFnVTable for CosineSimilarity {
     type Options = EmptyOptions;
+    type OptimizeVTable = ();
 
     fn id(&self) -> ScalarFnId {
         static ID: CachedId = CachedId::new("vortex.tensor.cosine_similarity");
